ENCD-3918 Add helper.py and remove helpers from search.py
